ARM: dts: am33xx: Add no-idle-on-init for lcdc node
In case of splash screen, u-boot configures the lcdc hw module
and displays the image.

But during kernel bootup, the ti-sysc driver disables the lcdc clock.
The lcdc driver configures the lcdc clock at a later stage.
During this period, the splash image is corrupted, and we see a glitch
on the display.

To fix this, add ti,no-idle-on-init property for the lcdc node so that
ti-sysc doesn't disable the clock, and the splash screen displayed by
the u-boot will be preserved till the lcdc driver takes over in the
kernel.

Signed-off-by: Sukrut Bellary <sbellary@baylibre.com>